refactor: useHandleMultiTasksの作成
ロジックを分離するため
